Each change about pull request #21
