+++ app/views/shopping/index.html.erb
@@ -37,6 +37,7 @@
         <% end %>
     </div>
     <br>
+    
     <br>
 </body>
 </html>
